+++ api/src/main/java/net/jqwik/api/statistics/StatisticsCoverage.java
@@ -12,7 +12,7 @@ import static org.apiguardian.api.API.Status.*;
  *
  * @see Statistics#coverage(Consumer)
  */
-@API(status = EXPERIMENTAL, since = "1.2.3")
+@API(status = MAINTAINED, since = "1.4.0")
 public interface StatisticsCoverage {
 
 	interface CoverageChecker {